I changed it so that It would adapt to backstabbers and loyal vengeful.
I made it so that if I get betrayed then I return 'b' to the
backstabbers so if they continue to backstab then I will also backstab.
against loyal vengeful as long as they continue to press 'c' then I will
also be loyal and continue to press 'c'. In summary I has lost the least
amount of points.